Controller refectored
