+++ TaskOrganizer/Gateway/TaskOrganizer.Api/Models/TaskModel.cs
@@ -2,9 +2,8 @@ using System;
 
 namespace TaskOrganizer.Api.Models
 {
-    public abstract class BaseTask
-    {
-        
+    public class TaskModel
+    {        
         public int TaskNumeber { get; set; }
         public string Title { get; set; }
         public string Description { get; set; }
