--- TaskOrganizer/Gateway/TaskOrganizer.Api/Models/TaskRequest.cs
@@ -1,7 +1,0 @@
-
-namespace TaskOrganizer.Api.Models
-{
-    public class TaskRequest : BaseTask
-    {
-    }
-}
