--- TaskOrganizer/Gateway/TaskOrganizer.Api/Models/TaskResponse.cs
@@ -1,7 +1,0 @@
-
-namespace TaskOrganizer.Api.Models
-{
-    public class TaskResponse : BaseTask
-    {
-    }
-}
